Split MCP parity gates
